backend/pyproject.toml
[tool.poetry]
name = "pressure-back"
version = "0.1.0"
description = ""
authors = ["Vladimir <justlikevova@gmail.com>"]
readme = "README.md"
packages = [{include = "pressure_back"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
drf-yasg = "^1.21.5"
flake8 = "^6.0.0"
django-cors-headers = "^3.14.0"
pre-commit = "^3.2.2"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.0.2"}
django-filter = "^23.1"
djoser = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
